♻️ Refactored code to be more react style
