+++ src/ts/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { createRoot } from 'react-dom/client';
 
-import { IconCards } from './card-container';
+import { App } from './App';
 
 const root = createRoot(document.querySelector('main'));
-root.render(<IconCards />);
+root.render(<App />);
